third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/transferBackFromGPUTexture-context-lost.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-promise-test.js"></script>
<script src="./webgpu-helpers.js"></script>

<h1>Canvas2D WebGPU Transfer: transferBackFromGPUTexture-context-lost</h1>
<p>transferBackFromGPUTexture() should be a no-op if the canvas context is lost.
</p>

<script>

runCanvasTestsInWorker({dependencies: ['./webgpu-helpers.js']});

/**
 * transferBackFromGPUTexture() should be a no-op if the canvas context is lost.
 */
canvasPromiseTest(async (canvas) => {
  const device = await getWebGPUDevice();

  // Begin a WebGPU access session.
  const ctx = canvas.getContext('2d');
  const tex = ctx.transferToGPUTexture({device: device});

  // Forcibly lose the canvas context.
  assert_true(!!self.internals, 'Internal APIs unavailable.');
  self.internals.forceLoseCanvasContext(ctx);
  assert_true(ctx.isContextLost());

  // End the WebGPU access session. Nothing should be thrown.
  try {
    ctx.transferBackFromGPUTexture();
  } catch {
    assert_unreached('transferBackFromGPUTexture should be safe when context ' +
                     'is lost.');
  }
}, 'transferBackFromGPUTexture() should be a no-op if the 2D context is lost.');

</script>
